Voronoi-LLoyd distributed
Now all robots are used for the tasselation, and their objective is set by their working state and their target. The lineup function ROB_lineup is replaced by a more "distributed" approach that uses a ring-pdf to guide robots to their lineup positions, while avoiding collisions. The sigma of each robot is controlled and adapted iteratively in order to provide safe indications for the robot (large sigma when distant for safety, small sigma when close for precision). Works well but not extremely well in case of crowded inbound zone. NOTE: please do not change sigmas randomly as they were precisely tuned.