Lineup -> Transportation -> Delivery check
Simulation now performs the 3 steps listed above, using various new functions and improvements.
A new working state was added to the voronoi to avoid robots being "bullied" in a corner by the formation. Delivery target for the package is chosen randomly inside one of the three outbound zones.
Various other changes in the simulation script to make the code ready for a multi-package simulation.